+++ src/Predictor.Models/DataStructures/Hourly/WindDirection.cs
@@ -3,7 +3,7 @@ using Microsoft.ML.Data;
 
 namespace Predictor.Models.DataStructures.Hourly
 {
-    public class WindDirection : Cities<double>
+    public class WindDirection : Cities<float>
     {
         [LoadColumn(0)]
         public DateTime DateTime { get; set; }
