+++ src/Predictor.Models/DataStructures/Hourly/WindSpeed.cs
@@ -3,7 +3,7 @@ using Microsoft.ML.Data;
 
 namespace Predictor.Models.DataStructures.Hourly
 {
-    public class WindSpeed : Cities<double>
+    public class WindSpeed : Cities<float>
     {
         [LoadColumn(0)]
         public DateTime DateTime { get; set; }
